requirements.txt
pandas==2.2.2
requests==2.31.0
sqlalchemy==2.0.30
streamlit==1.35.0
matplotlib==3.9.0
plotly==5.22.0
python-dotenv==1.0.1
loguru==0.7.2
